+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v0.3.2 (2024-06-22)
 
 ### Bug Fixes